Move six to foreign.six
