+++ opensvc/core/collector/cli.py
@@ -22,9 +22,9 @@ if TERM:
 import atexit
 import fnmatch
 
-from six.moves import configparser as ConfigParser
+from foreign.six.moves import configparser as ConfigParser
 import core.exceptions as ex
-from six.moves import input
+from foreign.six.moves import input
 from utilities.storage import Storage
 from utilities.proc import find_editor
 from utilities.render.color import formatter
