+++ opensvc/daemon/handlers/events/get.py
@@ -1,6 +1,6 @@
 import time
 
-from six.moves import queue
+from foreign.six.moves import queue
 import daemon.handler
 from env import Env
 from utilities.string import bdecode
