+++ opensvc/daemon/handlers/sync/get.py
@@ -1,6 +1,6 @@
 import time
 
-from six.moves import queue
+from foreign.six.moves import queue
 import daemon.handler
 import daemon.shared as shared
