+++ opensvc/drivers/array/nexenta.py
@@ -2,8 +2,8 @@ import json
 import base64
 
 import core.exceptions as ex
-from six.moves.urllib.request import Request, urlopen # pylint: disable=import-error
-from six.moves.urllib.error import URLError # pylint: disable=import-error
+from foreign.six.moves.urllib.request import Request, urlopen # pylint: disable=import-error
+from foreign.six.moves.urllib.error import URLError # pylint: disable=import-error
 from utilities.naming import factory, split_path
 from core.node import Node
