+++ opensvc/drivers/resource/task/__init__.py
@@ -6,7 +6,7 @@ import utilities.lock
 from core.scheduler import SchedOpts
 from utilities.lazy import lazy
 from core.resource import Resource
-from six.moves import input
+from foreign.six.moves import input
 from utilities.proc import lcall
 
 KEYWORDS = [
